phase 1 complete :wq :"

+++ db/migrate/20140107133451_create_users.rb
@@ -2,7 +2,7 @@ class CreateUsers < ActiveRecord::Migration
   def up
     create_table :users do |t|
       t.string :name
-      t.integer :leave
+      t.float :leave
 
       t.timestamps
     end
